refactor(web): simplify auth providers by extracting shared logic
