+++ apps/web/src/components/providers/CommunityAuthProvider.tsx
@@ -1,19 +1,17 @@
 import { createContext } from 'react';
 import { IOrganizationEntity, IUserEntity } from '@novu/shared';
-import { setUser as sentrySetUser, configureScope as sentryConfigureScope } from '@sentry/react';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQuery } from '@tanstack/react-query';
 import { HttpStatusCode } from 'axios';
-import { useLDClient } from 'launchdarkly-react-client-sdk';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
-import { ROUTES, PUBLIC_ROUTES_PREFIXES } from '../../constants/routes';
-import { useSegment } from './SegmentProvider';
+import { useCallback, useEffect, useState } from 'react';
+import { ROUTES } from '../../constants/routes';
 import { clearEnvironmentId } from './EnvironmentProvider';
 import { getToken } from '../../auth/getToken';
 import { getTokenClaims } from '../../auth/getTokenClaims';
 import { getUser } from '../../api/user';
 import { switchOrganization as apiSwitchOrganization, getOrganizations } from '../../api/organization';
 import { type AuthContextValue } from './AuthProvider';
+import { useCommonAuth } from '../../hooks/useCommonAuth';
+import { useNavigate, useLocation } from 'react-router-dom';
 
 // TODO: Add a novu prefix to the local storage key
 export const LOCAL_STORAGE_AUTH_TOKEN_KEY = 'auth_token';
@@ -26,8 +24,8 @@ export const CommunityAuthContext = createContext<AuthContextValue>({
   inPrivateRoute: false,
   isLoading: false,
   currentUser: null,
-  currentOrganization: null,
   organizations: [],
+  currentOrganization: null,
   login: asyncNoop,
   logout: noop,
   redirectToLogin: noop,
@@ -54,15 +52,15 @@ function inIframe() {
   }
 }
 
-function selectOrganization(organizations: IOrganizationEntity[] | null, selectedOrganizationId?: string) {
-  let org: IOrganizationEntity | null = null;
+function selectOrganization(organizations: IOrganizationEntity[] | undefined, selectedOrganizationId?: string) {
+  let org: IOrganizationEntity | undefined = undefined;
 
   if (!organizations) {
-    return null;
+    return;
   }
 
   if (selectedOrganizationId) {
-    org = organizations.find((currOrg) => currOrg._id === selectedOrganizationId) || null;
+    org = organizations.find((currOrg) => currOrg._id === selectedOrganizationId);
   }
 
   // Or pick the development environment
@@ -74,36 +72,14 @@ function selectOrganization(organizations: IOrganizationEntity[] | null, selecte
 }
 
 export const CommunityAuthProvider = ({ children }: { children: React.ReactNode }) => {
-  const ldClient = useLDClient();
-  const segment = useSegment();
-  const queryClient = useQueryClient();
   const navigate = useNavigate();
   const location = useLocation();
-  const inPublicRoute = !!Array.from(PUBLIC_ROUTES_PREFIXES.values()).find((prefix) =>
-    location.pathname.startsWith(prefix)
-  );
-  const inPrivateRoute = !inPublicRoute;
+
+  const [currentOrganization, setCurrentOrganization] = useState<IOrganizationEntity | undefined>(undefined);
   const hasToken = !!getToken();
 
-  useEffect(() => {
-    if (!getToken() && inPrivateRoute && !inIframe()) {
-      navigate(ROUTES.AUTH_LOGIN, { state: { redirectTo: location } });
-    }
-  }, [navigate, inPrivateRoute, location]);
-
-  const { data: user = null, isLoading: isUserLoading } = useQuery<IUserEntity>(['/v1/users/me'], getUser, {
-    enabled: hasToken,
-    retry: false,
-    staleTime: Infinity,
-    onError: (error: any) => {
-      if (error?.statusCode === HttpStatusCode.Unauthorized) {
-        logout();
-      }
-    },
-  });
-
   const {
-    data: organizations = null,
+    data: organizations,
     isLoading: isOrganizationLoading,
     refetch: refetchOrganizations,
   } = useQuery<IOrganizationEntity[]>(['/v1/organizations'], getOrganizations, {
@@ -117,16 +93,25 @@ export const CommunityAuthProvider = ({ children }: { children: React.ReactNode 
     },
   });
 
-  const reloadOrganization = async () => {
-    const { data } = await getOrganizations();
-    // we need to update all organizations so current org (data) and 'organizations' are not ouf of sync
-    await refetchOrganizations();
-    setCurrentOrganization(selectOrganization(data, currentOrganization?._id));
-  };
+  const { data: user, isLoading: isUserLoading } = useQuery<IUserEntity>(['/v1/users/me'], getUser, {
+    enabled: hasToken,
+    retry: false,
+    staleTime: Infinity,
+    onError: (error: any) => {
+      if (error?.statusCode === HttpStatusCode.Unauthorized) {
+        logout();
+      }
+    },
+  });
 
-  const [currentOrganization, setCurrentOrganization] = useState<IOrganizationEntity | null>(
-    selectOrganization(organizations)
-  );
+  const { inPublicRoute, inPrivateRoute, logout, redirectToLogin, redirectToSignUp } = useCommonAuth({
+    user,
+    organization: currentOrganization,
+    logoutCallback: async () => {
+      saveToken(null);
+      clearEnvironmentId();
+    },
+  });
 
   const login = useCallback(
     async (newToken: string, redirectUrl?: string) => {
@@ -146,63 +131,6 @@ export const CommunityAuthProvider = ({ children }: { children: React.ReactNode 
       redirectUrl ? navigate(redirectUrl) : void 0;
     },
     [navigate, refetchOrganizations]
-  );
-
-  const logout = useCallback(() => {
-    saveToken(null);
-    queryClient.clear();
-    segment.reset();
-    // TODO: Revise storing environment id in local storage to avoid having to clear it during org or env switching
-    clearEnvironmentId();
-    navigate(ROUTES.AUTH_LOGIN);
-    setCurrentOrganization(null);
-  }, [navigate, queryClient, segment]);
-
-  const redirectTo = useCallback(
-    ({
-      url,
-      redirectURL,
-      origin,
-      anonymousId,
-    }: {
-      url: string;
-      redirectURL?: string;
-      origin?: string;
-      anonymousId?: string | null;
-    }) => {
-      const finalURL = new URL(url, window.location.origin);
-
-      if (redirectURL) {
-        finalURL.searchParams.append('redirect_url', redirectURL);
-      }
-
-      if (origin) {
-        finalURL.searchParams.append('origin', origin);
-      }
-
-      if (anonymousId) {
-        finalURL.searchParams.append('anonymous_id', anonymousId);
-      }
-
-      // Note: Do not use react-router-dom. The version we have doesn't do instant cross origin redirects.
-      window.location.replace(finalURL.href);
-    },
-    []
-  );
-
-  const redirectToLogin = useCallback(
-    ({ redirectURL }: { redirectURL?: string } = {}) => redirectTo({ url: ROUTES.AUTH_LOGIN, redirectURL }),
-    [redirectTo]
-  );
-
-  const redirectToSignUp = useCallback(
-    ({
-      redirectURL,
-      origin,
-      anonymousId,
-    }: { redirectURL?: string; origin?: string; anonymousId?: string | null } = {}) =>
-      redirectTo({ url: ROUTES.AUTH_SIGNUP, redirectURL, origin, anonymousId }),
-    [redirectTo]
   );
 
   const switchOrganization = useCallback(
@@ -225,6 +153,13 @@ export const CommunityAuthProvider = ({ children }: { children: React.ReactNode 
     [organizations, currentOrganization, setCurrentOrganization, login]
   );
 
+  const reloadOrganization = async () => {
+    const { data } = await getOrganizations();
+    // we need to update all organizations so current org (data) and 'organizations' are not ouf of sync
+    await refetchOrganizations();
+    setCurrentOrganization(selectOrganization(data, currentOrganization?._id));
+  };
+
   useEffect(() => {
     if (organizations) {
       setCurrentOrganization(selectOrganization(organizations, getTokenClaims()?.organizationId));
@@ -232,40 +167,10 @@ export const CommunityAuthProvider = ({ children }: { children: React.ReactNode 
   }, [organizations, currentOrganization, switchOrganization]);
 
   useEffect(() => {
-    if (user && currentOrganization) {
-      segment.identify(user);
-
-      sentrySetUser({
-        email: user.email ?? '',
-        username: `${user.firstName} ${user.lastName}`,
-        id: user._id,
-        organizationId: currentOrganization._id,
-        organizationName: currentOrganization.name,
-      });
-    } else {
-      sentryConfigureScope((scope) => scope.setUser(null));
+    if (!getToken() && inPrivateRoute && !inIframe()) {
+      navigate(ROUTES.AUTH_LOGIN, { state: { redirectTo: location } });
     }
-  }, [user, currentOrganization, segment]);
-
-  useEffect(() => {
-    if (!ldClient) {
-      return;
-    }
-
-    if (currentOrganization) {
-      ldClient.identify({
-        kind: 'organization',
-        key: currentOrganization._id,
-        name: currentOrganization.name,
-        createdAt: currentOrganization.createdAt,
-      });
-    } else {
-      ldClient.identify({
-        kind: 'user',
-        anonymous: true,
-      });
-    }
-  }, [ldClient, currentOrganization]);
+  }, [navigate, inPrivateRoute, location]);
 
   const value = {
     inPublicRoute,
